requirements.txt
python-dotenv~=1.0.0
fastapi~=0.104.1
pymongo~=4.6.0
pydantic~=2.5.1
requests~=2.31.0